jQuery.Model class: Do not alias 'model' to 'wrap' and 'models' to 'wrapMany' in the prototype section of the class definition.
Instead, warn about deprecation of old methods and pass control to the new methods, honoring overridden methods as needed.

We cannot alias 'model' to 'wrap' in the prototype, because newer models will override 'model' and 'models' in the subclass rather than 'wrap' and 'wrapMany'. If we alias them e.g. $.Model.wrapMany = $.Model.models then whenever any legacy code calls 'wrapMany', it will bypass the overridden 'models' function and instead call 'models' on the base class, which is not the desired behavior.